+++ app/hooks/useDarkMode.ts
@@ -1,0 +1,26 @@
+// useDarkMode.ts
+import { useEffect, useState } from "react";
+
+export const useDarkMode = () => {
+  const [darkMode, setDarkMode] = useState<boolean>(true);
+
+  useEffect(() => {
+    let savedMode = localStorage.getItem("displayMode");
+    if (!savedMode) {
+      savedMode = "light";
+      setDarkMode(false);
+      localStorage.setItem("displayMode", savedMode);
+    }
+    setDarkMode(savedMode === "dark");
+  }, []);
+
+  const toggleDarkMode = () => {
+    setDarkMode((prev) => {
+      const newMode = !prev;
+      localStorage.setItem("displayMode", newMode ? "dark" : "light");
+      return newMode;
+    });
+  };
+
+  return { darkMode, toggleDarkMode };
+};
